requirements.txt
youtube-search-python 
youtube-transcript-api 
openai
httpx<0.24
yt-dlp
ffmpeg
ffmpeg-python
fastapi
python-dotenv
praw
pymongo
google-genai